+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 * Require `typing_extensions` on Python 3.11 (already required on earlier versinons) for better compatibility with pydantic v2
+* Fix `RawSimulator` handling of `cache_control` parameter during tests.
 
 ## [1.22.1] - 2023-07-24